+++ src/CommonRuntime/TextConversions.fs
@@ -38,20 +38,32 @@ module private Helpers =
 /// Conversions from string to string/int/int64/decimal/float/boolean/datetime/guid options
 type TextConversions = 
 
-  static member DefaultMissingValues = [|"NaN"; "NA"; "#N/A"; ":"; "-"; "TBA"; "TBD"|]
+  static member DefaultMissingValues = [| "NaN"; "NA"; "#N/A"; ":"; "-"; "TBA"; "TBD" |]
+  static member DefaultNonCurrencyAdorners = [| '%'; '‰'; '‱' |] |> Set.ofArray
+  static member DefaultCurrencyAdorners = [| '¤'; '$'; '¢'; '£'; '¥'; '₱'; '﷼'; '₤'; '₭'; '₦'; '₨'; '₩'; '₮'; '€'; '฿'; '₡'; '៛'; '؋'; '₴'; '₪'; '₫'; '₹'; 'ƒ' |] |> Set.ofArray
+
+  static member DefaultRemovableAdornerCharacters = 
+    Set.union TextConversions.DefaultNonCurrencyAdorners TextConversions.DefaultCurrencyAdorners
+  
+  //This removes any adorners that might otherwise casue the inference to infer string. A notable a change is
+  //Currency Symbols are now treated as an Adorner like a '%' sign thus are now independant
+  //of the culture. Which is probably better since we have lots of scenarios where we want to
+  //consume values prefixed with € or $ but in a different culture. 
+  static member RemoveAdorners (value:string) = 
+      new String(value.ToCharArray() |> Array.filter (not << TextConversions.DefaultRemovableAdornerCharacters.Contains))
 
   /// Turns empty or null string value into None, otherwise returns Some
   static member AsString str =
     if String.IsNullOrWhiteSpace str then None else Some str
 
   static member AsInteger cultureInfo text = 
-    Int32.TryParse(text, NumberStyles.Integer, cultureInfo) |> asOption
+    Int32.TryParse(TextConversions.RemoveAdorners text, NumberStyles.Integer, cultureInfo) |> asOption
   
   static member AsInteger64 cultureInfo text = 
-    Int64.TryParse(text, NumberStyles.Integer, cultureInfo) |> asOption
+    Int64.TryParse(TextConversions.RemoveAdorners text, NumberStyles.Integer, cultureInfo) |> asOption
   
   static member AsDecimal cultureInfo text =
-    Decimal.TryParse(text, NumberStyles.Number ||| NumberStyles.AllowCurrencySymbol, cultureInfo) |> asOption
+    Decimal.TryParse(TextConversions.RemoveAdorners text, NumberStyles.Number, cultureInfo) |> asOption
   
   /// if useNoneForMissingValues is true, NAs are returned as None, otherwise Some Double.NaN is used
   static member AsFloat missingValues useNoneForMissingValues cultureInfo (text:string) =