+++ tests/FSharp.Data.Tests/HtmlParser.fs
@@ -1,4 +1,5 @@
-﻿#if INTERACTIVE
+﻿
+#if INTERACTIVE
 #r "../../bin/FSharp.Data.dll"
 #r "../../packages/NUnit.2.6.3/lib/nunit.framework.dll"
 #load "../Common/FsUnit.fs"
@@ -38,30 +39,32 @@ let ``Can handle unclosed tags correctly``() =
                         </body>
                     </html>"""
     let result = HtmlDocument.Parse simpleHtml
+    
     let expected = 
-        HtmlDocument.New
-         [
-           HtmlNode.NewElement("html",
-            [
-               HtmlNode.NewElement("head",
-                [
-                    HtmlNode.NewElement("script", ["language","JavaScript"
-                                                   "src","/bwx_generic.js"])
-                    HtmlNode.NewElement("link", ["rel","stylesheet"
-                                                 "type","text/css"
-                                                 "href","/bwx_style.css"])
-                ])
-               HtmlNode.NewElement("body",
-                [
-                    HtmlNode.NewElement("img", ["src", "myimg.jpg"])
-                    HtmlNode.NewElement("table", ["title", "table"],
-                     [
-                        HtmlNode.NewElement("tr", [HtmlNode.NewElement("th", [HtmlNode.NewText "Column 1"]); HtmlNode.NewElement("th", [HtmlNode.NewText "Column 2"])])
-                        HtmlNode.NewElement("tr", [HtmlNode.NewElement("td", [HtmlNode.NewText "1"]); HtmlNode.NewElement("td", [HtmlNode.NewText "yes"])])
-                     ])
-                ])
-            ])
-        ]
+        HtmlDocument.New 
+            [ HtmlNode.NewElement
+                  ("html", 
+                   [ HtmlNode.NewElement("head", 
+                                         [ HtmlNode.NewElement("script", 
+                                                               [ "language", "JavaScript"
+                                                                 "src", "/bwx_generic.js" ])
+                                           HtmlNode.NewElement("link", 
+                                                               [ "rel", "stylesheet"
+                                                                 "type", "text/css"
+                                                                 "href", "/bwx_style.css" ]) ])
+                     
+                     HtmlNode.NewElement
+                         ("body", 
+                          [ HtmlNode.NewElement("img", [ "src", "myimg.jpg" ])
+                            
+                            HtmlNode.NewElement
+                                ("table", [ "title", "table" ], 
+                                 [ HtmlNode.NewElement("tr", 
+                                                       [ HtmlNode.NewElement("th", [ HtmlNode.NewText "Column 1" ])
+                                                         HtmlNode.NewElement("th", [ HtmlNode.NewText "Column 2" ]) ])
+                                   HtmlNode.NewElement("tr", 
+                                                       [ HtmlNode.NewElement("td", [ HtmlNode.NewText "1" ])
+                                                         HtmlNode.NewElement("td", [ HtmlNode.NewText "yes" ]) ]) ]) ]) ]) ]
     result |> should equal expected
 
 [<Test>]
@@ -69,6 +72,17 @@ let ``Can handle multiple char refs in a text run``() =
     let html = HtmlNode.Parse "<div>&quot;Foo&quot;</div>"
     let result = html.Head.InnerText()
     result |> should equal "\"Foo\""
+
+[<Test>]
+let ``Can handle attributes with no value``() = 
+    let html = """<li itemscope itemtype="http://schema.org/Place"></li>"""
+    let node = HtmlNode.Parse html |> List.head
+    let expected = 
+        [
+            HtmlAttribute.New("itemscope", "")
+            HtmlAttribute.New("itemtype", "http://schema.org/Place")
+        ]
+    node.Attributes() |> should equal expected
 
 [<Test>]
 let ``Can parse tables from a simple html``() = 
@@ -80,12 +94,16 @@ let ``Can parse tables from a simple html``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].HasHeaders |> should equal (Some true)
     tables.[0].Name |> should equal "table"
-    tables.[0].Rows |> should equal [["Column 1"]; ["1"]]
+    tables.[0].Rows |> should equal [ [ "Column 1" ]
+                                      [ "1" ] ]
 
 [<Test>]
 let ``Can parse tables from a simple html table but infer headers``() = 
@@ -98,12 +116,17 @@ let ``Can parse tables from a simple html table but infer headers``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].HasHeaders |> should equal (Some true)
     tables.[0].Name |> should equal "table"
-    tables.[0].Rows |> should equal [["Column 1"]; ["1"]; ["2"]]
+    tables.[0].Rows |> should equal [ [ "Column 1" ]
+                                      [ "1" ]
+                                      [ "2" ] ]
 
 [<Test>]
 let ``Ignores empty tables``() = 
@@ -113,8 +136,11 @@ let ``Ignores empty tables``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 0
 
 [<Test>]
@@ -128,12 +154,17 @@ let ``Can parse tables with no headers``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "table"
     tables.[0].HasHeaders |> should equal (Some false)
-    tables.[0].Rows |> should equal [["2"]; ["1"]; ["3"]]
+    tables.[0].Rows |> should equal [ [ "2" ]
+                                      [ "1" ]
+                                      [ "3" ] ]
 
 [<Test>]
 let ``Can parse tables with no headers and only 2 rows``() = 
@@ -145,12 +176,16 @@ let ``Can parse tables with no headers and only 2 rows``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "table"
     tables.[0].HasHeaders |> should equal (Some false)
-    tables.[0].Rows |> should equal [["1"]; ["3"]]
+    tables.[0].Rows |> should equal [ [ "1" ]
+                                      [ "3" ] ]
 
 [<Test>]
 let ``Extracts table when title attribute is set``() = 
@@ -162,8 +197,11 @@ let ``Extracts table when title attribute is set``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "table"
 
@@ -177,8 +215,11 @@ let ``Extracts table when name attribute is set``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "table"
 
@@ -192,8 +233,11 @@ let ``When mutiple identifying attributes are set the id attribute is selected``
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "table_id"
 
@@ -207,8 +251,11 @@ let ``When mutiple identifying attributes are set but not the id attribute is th
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "table_name"
 
@@ -222,7 +269,11 @@ let ``Extracts tables without an id title or name attribute``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
 
 [<Test>]
@@ -247,11 +298,18 @@ let ``Extracts data and headers with thead and tbody``() =
                       </tr>
                     </tbody>
                   </table>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "savings_table"
     tables.[0].HasHeaders |> should equal (Some true)
-    tables.[0].Rows |> should equal [["Month";"Savings"]; ["Sum"; "$180"]; ["January"; "$100"]; ["February"; "$80"]]
+    tables.[0].Rows |> should equal [ [ "Month"; "Savings" ]
+                                      [ "Sum"; "$180" ]
+                                      [ "January"; "$100" ]
+                                      [ "February"; "$80" ] ]
 
 [<Test>]
 let ``Extracts tables in malformed html``() = 
@@ -264,33 +322,41 @@ let ``Extracts tables in malformed html``() =
                         </table>
                     </body>
                 </html>"""
-    let tables = html |> HtmlDocument.Parse |> getTables true
-
+    
+    let tables = 
+        html
+        |> HtmlDocument.Parse
+        |> getTables true
     tables.Length |> should equal 1
     tables.[0].Name |> should equal "Table1"
     tables.[0].HasHeaders |> should equal (Some true)
-    tables.[0].Rows |> should equal [["Column 1"]; ["1"]]
-
+    tables.[0].Rows |> should equal [ [ "Column 1" ]
+                                      [ "1" ] ]
 
 [<Test>]
 let ``Can handle html with doctype and xml namespaces``() = 
-    let html = """<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd"><html lang="en" xml:lang="en" xmlns="http://www.w3.org/1999/xhtml"><body>content</body></html>"""
+    let html = 
+        """<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd"><html lang="en" xml:lang="en" xmlns="http://www.w3.org/1999/xhtml"><body>content</body></html>"""
     let htmlDoc = HtmlDocument.Parse html
+    
     let expected = 
-            HtmlDocument.New("html PUBLIC \"-//W3C//DTD XHTML 1.0 Strict//EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd\"",
-             [
-               HtmlNode.NewElement("html", ["lang","en"; "xml:lang","en"; "xmlns","http://www.w3.org/1999/xhtml"], [HtmlNode.NewElement("body", [HtmlNode.NewText "content"])])
-             ])
+        HtmlDocument.New
+            ("html PUBLIC \"-//W3C//DTD XHTML 1.0 Strict//EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd\"", 
+             [ HtmlNode.NewElement
+                   ("html", 
+                    [ "lang", "en"
+                      "xml:lang", "en"
+                      "xmlns", "http://www.w3.org/1999/xhtml" ], 
+                    [ HtmlNode.NewElement("body", [ HtmlNode.NewText "content" ]) ]) ])
     expected |> should equal htmlDoc
 
 [<Test>]
 let ``Can find header when nested in a div``() = 
     let tables = 
-        HtmlDocument.Load "data/wimbledon_wikipedia.html" 
+        HtmlDocument.Load "data/wimbledon_wikipedia.html"
         |> getTables false
         |> List.map (fun t -> t.Name, t)
         |> Map.ofList
-    
     Map.containsKey "Ranking points" tables |> should equal true
     Map.containsKey "Records" tables |> should equal true
     Map.containsKey "Current champions" tables |> should equal true
@@ -299,7 +365,7 @@ let ``Can find header when nested in a div``() =
 let ``Can parse tables imdb chart``() = 
     let imdb = HtmlDocument.Load "data/imdb_chart.htm"
     let tables = imdb |> getTables false
-    tables.Length |> should equal 1
+    tables.Length |> should equal 2
     tables.[0].Name |> should equal "Top 250"
     tables.[0].HasHeaders |> should equal (Some true)
     tables.[0].Rows.Length |> should equal 251
@@ -310,19 +376,19 @@ let ``Can parse tables ebay cars``() =
     true |> should equal true
 
 [<Test>]
-let ``Does not crash when parsing us presidents``() =
+let ``Does not crash when parsing us presidents``() = 
     let table = HtmlDocument.Load "data/us_presidents_wikipedia.html" |> getTables false
     true |> should equal true
 
 [<Test>]
-let ``Can parse non-self-closing tags of elements that can't have children when followed by comments``() =
+let ``Can parse non-self-closing tags of elements that can't have children when followed by comments``() = 
     let html = """<hr class="hr4"><!--comment1--><!--comment2--><hr class="hr5" />"""
     let expected = """<hr class="hr4" /><!--comment1--><!--comment2--><hr class="hr5" />"""
     let result = (HtmlDocument.Parse html).ToString()
     result |> shouldEqual expected
 
 [<Test>]
-let ``Ignores spurious closing tags``() =
+let ``Ignores spurious closing tags``() = 
     let html = 
         """<li class="even"><a class="clr" href="/pj/ldbdetails/kEW6eAApVxWdogIXhoHAew%3D%3D/?board=dep"><span class="time em">21:36<br /><small>On time</small></span></span><span class="station">Brighton (East Sussex)</span><span class="platform"><small>Platform</small><br />17</span></a></li>"""
     let expected = """<li class="even">
@@ -338,27 +404,117 @@ let ``Ignores spurious closing tags``() =
   </a>
 </li>"""
     let result = (HtmlDocument.Parse html).ToString().Replace("\r", null)
-    result |> shouldEqual <| expected.Replace("\r", null)
+    result
+    |> shouldEqual
+    <| expected.Replace("\r", null)
 
 [<Test>]
-let ``Can parse national rail mobile site correctly``() =
-    HtmlDocument.Load "data/UKDepartures.html" |> HtmlDocument.descendantsNamed false ["li"] |> List.length |> should equal 68
-    HtmlDocument.Load "data/UKLiveProgress.html" |> HtmlDocument.descendantsNamed false ["li"] |> List.length |> should equal 15
-    HtmlDocument.Load "data/UKDepartures.html" |> HtmlDocument.descendantsNamed false ["li"; "hr"] |> List.length |> should equal 69
-    HtmlDocument.Load "data/UKLiveProgress.html" |> HtmlDocument.descendantsNamed false ["li"; "hr"] |> List.length |> should equal 17
+let ``Can handle CDATA blocks``() = 
+    let html = """
+    <!DOCTYPE html>
+    <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en" xmlns:fb="http://www.facebook.com/2008/fbml" xmlns:og="http://opengraphprotocol.org/schema/">
+     <head>
+        <script type="text/javascript">
+            var google_tag_params = { PROP_intent: "RENT", PROP_use: "RES", PROP_loc: "London", PROP_minprice: "1500", PROP_maxprice: "1750", PROP_beds: "1" };
+        </script>
+
+        <script type="text/javascript">
+         <![CDATA[
+            var google_conversion_id = 964294565;
+            var google_custom_params = window.google_tag_params;
+            var google_remarketing_only = true;
+         ]]>
+        </script>
+     </head>
+     <body>
+         <ul>
+             <li>1</li>
+             <li>2</li>
+         <ul>
+     </body>
+    </html>
+    """
+    
+    let doc = HtmlDocument.Parse html
+    let result = 
+        doc
+        |> HtmlDocument.descendantsNamed false [ "li" ]
+        |> List.map (HtmlNode.innerText)
+    result |> should equal [ "1"; "2"]
 
 [<Test>]
-let ``Can parse zoopla site correctly``() =
-    HtmlDocument.Load "data/zoopla.html" 
-    |> HtmlDocument.descendantsNamed false ["li"]
+let ``Can parse nested lists correctly when stops on recurse``() = 
+    let html = """
+        <ul>
+            <li>
+                <ul>
+                    <li>1</li>
+                    <li>2</li>
+                </ul>
+            </li>
+            <li>3</li>
+            <li>4</li>
+       </ul>
+    """
+    
+    let result = 
+        (HtmlDocument.Parse html)
+        |> HtmlDocument.descendantsNamed false [ "li" ]
+        |> List.map (HtmlNode.innerText)
+    result |> should equal [ "12"; "3"; "4" ]
+
+[<Test>]
+let ``Can parse nested lists correctly when continues on recurse``() = 
+    let html = """
+        <ul>
+            <li>
+                <ul>
+                    <li>1</li>
+                    <li>2</li>
+                </ul>
+            </li>
+            <li>3</li>
+            <li>4</li>
+       </ul>
+    """
+    
+    let result = 
+        (HtmlDocument.Parse html)
+        |> HtmlDocument.descendantsNamed true [ "li" ]
+        |> List.map (HtmlNode.innerText)
+    result |> should equal [ "12"; "1"; "2"; "3"; "4" ]
+
+[<Test>]
+let ``Can parse national rail mobile site correctly``() = 
+    HtmlDocument.Load "data/UKDepartures.html"
+    |> HtmlDocument.descendantsNamed false [ "li" ]
+    |> List.length
+    |> should equal 68
+    HtmlDocument.Load "data/UKLiveProgress.html"
+    |> HtmlDocument.descendantsNamed false [ "li" ]
+    |> List.length
+    |> should equal 15
+    HtmlDocument.Load "data/UKDepartures.html"
+    |> HtmlDocument.descendantsNamed false [ "li"; "hr" ]
+    |> List.length
+    |> should equal 69
+    HtmlDocument.Load "data/UKLiveProgress.html"
+    |> HtmlDocument.descendantsNamed false [ "li"; "hr" ]
+    |> List.length
+    |> should equal 17
+
+[<Test>]
+let ``Can parse zoopla site correctly``() = 
+    HtmlDocument.Load "data/zoopla.html"
+    |> HtmlDocument.descendantsNamed true ["li"]
     |> List.filter (HtmlNode.hasAttribute "itemtype" "http://schema.org/Place")
     |> List.length 
     |> should equal 100
 
 [<Test>]
-let ``Doesn't insert whitespace on attribute name when there are two whitespace characters before an attribute``() =
-    HtmlNode.Parse "<a data-lecture-id=\"27\"\r\ndata-modal-iframe=\"https://class.coursera.org/mathematicalmethods-001/lecture/view?lecture_id=27\"></a>"
+let ``Doesn't insert whitespace on attribute name when there are two whitespace characters before an attribute``() = 
+    HtmlNode.Parse 
+        "<a data-lecture-id=\"27\"\r\ndata-modal-iframe=\"https://class.coursera.org/mathematicalmethods-001/lecture/view?lecture_id=27\"></a>"
     |> List.head
     |> HtmlNode.attributeValue "data-modal-iframe"
     |> should equal "https://class.coursera.org/mathematicalmethods-001/lecture/view?lecture_id=27"
-